Fixed bugs 2744 (entities_in_radius using only 1 chunk) and 2760 (pitch as a float)
http://youtrack.sk89q.com/issue/CMDHELPER-2744
http://youtrack.sk89q.com/issue/CMDHELPER-2760

+++ src/main/java/com/laytonsmith/abstraction/MCPlayer.java
@@ -115,7 +115,7 @@ public interface MCPlayer extends MCCommandSender, MCHumanEntity,
 
 	public void playNote(MCLocation loc, MCInstrument instrument, MCNote note);
 	
-	public void playSound(MCLocation l, MCSound sound, int volume, int pitch);
+	public void playSound(MCLocation l, MCSound sound, float volume, float pitch);
 
 	public int getHunger();
