+++ src/main/java/com/laytonsmith/abstraction/MCWorld.java
@@ -42,7 +42,7 @@ public interface MCWorld extends AbstractionObject{
 
     public void playEffect(MCLocation l, MCEffect mCEffect, int e, int data);
 
-	public void playSound(MCLocation l, MCSound sound, int volume, int pitch);
+	public void playSound(MCLocation l, MCSound sound, float volume, float pitch);
 
     public MCItem dropItemNaturally(MCLocation l, MCItemStack is);
